% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{checkers}
\alias{checkers}
\alias{check_linewidth}
\alias{check_comments}
\title{Functions to check coding standards}
\usage{
check_linewidth(obj, n = 90)

check_comments(obj)
}
\arguments{
\item{obj}{An object of class \link{sdb_files}}

\item{n}{Threshold for max linewidth.}
}
\description{
Functions to check coding standards
}
\examples{

\dontrun{
# Looking at the files in the project
obj <- read_files()

# Calling the checkers
check_comments(obj)
check_linewidth(obj)
}


}
