% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lints.R
\name{comments_linter}
\alias{comments_linter}
\title{A lint to count proportion of code/comments}
\usage{
comments_linter(source_file)
}
\arguments{
\item{source_file}{See \link[lintr:linters]{lintr::linters}}
}
\description{
A lint to count proportion of code/comments
}
