% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{is_rpackage}
\alias{is_rpackage}
\title{Check whether a path is an R package or not.}
\usage{
is_rpackage(path = ".", fn = c("R", "DESCRIPTION", "NAMESPACE", "man"))
}
\arguments{
\item{path}{Character scalar. Path to the package to dir.}

\item{fn}{Character vector. List of files to consider an R package.}
}
\description{
Check whether a path is an R package or not.
}
\examples{
is_rpackage()
}
