% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-class.R
\name{read_files}
\alias{read_files}
\alias{sdb_files}
\title{Read-in files to be processed by the checkers}
\usage{
read_files(path = ".", trim = TRUE, ext = c(R = "[.](R|r)$", Rmd =
  "[.](Rmd|rmd)$", cpp = "[.]cpp$", h = "[.]h(pp)?$"))
}
\arguments{
\item{path}{Path to the project to process data from}

\item{trim}{Logical scalar. When \code{TRUE}, will remove ending blanks.}

\item{ext}{Regular expression to capture in filenames.}
}
\description{
Read-in files to be processed by the checkers
}
